refactored fetch function
